// fe_inventarios_gei/src/app/models/usuario.ts
export interface IUsuario {
    sDistrito?: String;
    sProvincia?: String;
    sDepartamento?: String;
    sUsuario?: String;
    sNombres?: String;
    sApellidos?: String;
    nIdInstitucion?: Number;
    sNombreInstitucion?: String;
    nIdRol?: Number;
    sToken?: String;
    nTipo?: String;
    sTipoMunicipalidad?: any;
}